// final/src/app/model/bds.ts
export  class Bds {
  id: number;
  danhMuc: string;
  vungMien: string;
  banLa: string;
  banDangTin: string;
  tinhTrang: string;
  diachi: string;
  dienTich: number;
  huong: string;
  tuaDe: string;
  noiDung:string;
  gia: number;
  ngayDang: string;


  constructor() {
    this.id = 0;
    this.danhMuc = "";
    this.vungMien = "";
    this.banLa = "";
    this.banDangTin = "";
    this.tinhTrang = "";
    this.diachi = "";
    this.dienTich = 0;
    this.huong = "";
    this.tuaDe = "";
    this.noiDung = "";
    this.gia = 0;
    this.ngayDang="";
  }
}